# Don't get a stack trace for every missing result in FindMissing