@tylerwilliams 2025-04-14
I think a better fix here would be to:
- define a flag alias in status.go that points at the app.log_error_stack_traces flag from log.go
- disable stack trace collection if that flag is set to false