@vanja-p 2025-04-14
Would you do that for all errors, or just NotFound?