@tylerwilliams 2025-04-14
Previously, we did not collect stack traces with errors. We added them with that logging flag, which I don't think is often used. I don't believe any code depends on them being set and if it did it shouldn't.