@vanja-p 2025-04-14
I used the flag to disable stack collection when stack logging is disabled.